Bug 910446: Handle duration changes in Music app
With the landing of bug 831224, the duration of MP3 files is
computed while they are playing, and updated accordingly. If
this happens, we distribute a 'durationchange' event to the
related apps.

With this patch, we handle 'durationchange' in the Music app.
All we have to do is to update the displayed duration and ensure
that the end timeout triggers at the end of the track.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>